refactor: consolidate e2e tests into focused test suites
Replace fragmented test files with two comprehensive test suites:
- full_stack_dev_test.go: Complete cluster lifecycle testing
- ha_operations_test.go: High availability and scaling operations

Remove redundant test files that duplicated coverage:
- backup_test.go, monitoring_test.go, selfhealing_test.go
- operator_centric_test.go, operator_scale_test.go
- full_stack_test.go, sequential_test.go

Update suite_test.go and operator_helpers.go to support new structure.
