+++ tests/e2e/suite_test.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"sync"
 	"testing"
 	"time"
 
@@ -14,6 +15,32 @@ import (
 	hcloud_client "github.com/imamik/k8zner/internal/platform/hcloud"
 	"github.com/imamik/k8zner/internal/provisioning/image"
 )
+
+// Test result tracking for skip logic between tests
+var (
+	testResultsLock sync.Mutex
+	fullStackPassed bool
+)
+
+// SetFullStackPassed marks the full stack test as passed.
+// This is called by TestE2EFullStackDev after successful completion.
+func SetFullStackPassed() {
+	testResultsLock.Lock()
+	defer testResultsLock.Unlock()
+	fullStackPassed = true
+	os.Setenv("E2E_FULLSTACK_PASSED", "true")
+}
+
+// IsFullStackPassed returns whether TestE2EFullStackDev passed.
+func IsFullStackPassed() bool {
+	// First check env var (allows external override)
+	if os.Getenv("E2E_FULLSTACK_PASSED") == "true" {
+		return true
+	}
+	testResultsLock.Lock()
+	defer testResultsLock.Unlock()
+	return fullStackPassed
+}
 
 // TestMain orchestrates E2E tests to run sequentially and manage shared resources.
 func TestMain(m *testing.M) {
